article-api/pyproject.toml
[tool.poetry]
name = "article-api"
version = "0.1.0"
description = ""
authors = ["Sebastiaan Gerritsen <s.e.gerritsen@outlook.com>"]
readme = "README.md"
packages = [{include = "article_api"}]

[tool.poetry.dependencies]
python = "^3.11"
haystack-ai = "^2.0.0b8"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-dotenv = "^1.0.1"
chroma-haystack = "^0.15.0"
pypdf = "^4.1.0"
ollama-haystack = "^0.0.5"
google-cloud-firestore = "^2.15.0"
firebase-admin = "^6.4.0"
dataclasses-json = "^0.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
